refactoring: removed abstract types

+++ src/utils/game/geometry.c
@@ -6,13 +6,13 @@
 /*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/05 11:50:00 by mmarinel          #+#    #+#             */
-/*   Updated: 2022/11/06 18:47:13 by mmarinel         ###   ########.fr       */
+/*   Updated: 2022/11/12 11:46:29 by mmarinel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "game_utils.h"
 
-t_2d_point	ft_rotate(t_2d_point vector, double ray_angle)
+t_2d_point	ft_rotate(t_2d_point vector, float ray_angle)
 {
 	const t_2d_matrix	rotation_matrix = ft_get_new_2dmtrx(
 			ft_get_new_2dpt(cos(ray_angle), sin(ray_angle)),
